+++ AudioVisio/src/audiovisio/level/Trigger.java
@@ -13,10 +13,10 @@ public class Trigger {
 
 	public static Trigger load(JSONObject obj) {
 		JSONObject location = (JSONObject) obj.get("location");
-		float locX = (float) location.get("x");
-		float locY = (float) location.get("y");
-		float locZ = (float) location.get("z");
-
+		float locX = ((Double)location.get("x")).floatValue();
+		float locY = ((Double)location.get("y")).floatValue();
+		float locZ = ((Double)location.get("z")).floatValue();
+		
 		return new Trigger(new Vector3f(locX, locY, locZ));
 	}
